Fix turning behavior for closing cross track error
Address differences in turning requirements needed to minimize cross track
error versus minimizing error when approaching a target object (cone).

We want the robot to always move in a forward direction (no turning on axis)
when minimizing cross track error. But when driving toward a cone, we want to
be able to turn on axis and go slower when approaching the cone.

Also address the object avoidance debounce counter behavior when intermittently
triggering the obstacle avoidance thresholds. Now the counter counts down twice
as quickly as it counts back up. So if the thresholds are regularly triggered,
the robot will eventually go into obstacle avoidance mode.