Turn back towards target when it leaves the field of view
Solve issue when the robot is moving toward a cone, with the cone in view, if
the cone goes out of view on the left or right of the frame, the robot does not
necessarily turn back toward the cone, but instead goes into whatever current
search pattern is selected.

Instead of reverting to an arbitrary search pattern, the robot goes into a
search pattern that will steer it back toward the cone. So when the cone leaves
the left hand side of the field of view. The robot goes into search pattern 2
where the robot turns left. And if the cone leaves the right side of the frame
the robot goes into pattern 3 where it turns right.