Increase cone search distance. Fix search pattern turning.
Increase target threshold distance from 6 meters to 8. Also increase the search
distance for each search pattern to 8 meters.

Decrease the turn rate (magnitude) during left and right search patterns to 10
(was 15). This ensures the the robot will move forward while turning rather than
turning in place.

Add a timeout timer to stop the robot for half a second when it first sees a
cone. This helps ensure the robot will approach the cone in a more controlled way.

Add magnetic declination for Almeda fair grounds. Disable IR rangefinders, the
grass at the grounds has not been cut recently and is so tall it interferes
with the IR sensors.

We tried to calibrate the encoder counts per meter given our new wheel diameter
of 49 cm. We arrived at a figure: 3052.406 counts per meter, but the robot
seems to overshoot targets with that value, so we decreased the counts per meter
to 2800, which works well.